* defoption.scm: Added. A pretty complete `define-scwm-option' implementation.  It might need revising for scheme style, and has lots of potential for further improvements, but is already pretty useful. Partially inspired by Emacs's customize.
* preferences.scm: `gui-set', `option-menu', `popup-option-menu',
ans `scwm-options-menu'.  Mostly replaces the functionality in
prefs-menu module, but does it using the information gleaned from
define-scwm-option invocations of loaded modules.

* primopts.scm: Added a handful of `define-scwm-option' calls for
options that are managed by primitives and not real variables.
Also added for a test boolean flag.  (A test string option is in
themes.scm)


(more later... but last bus is now!)