* window.c: Doc fix to get-window-highlight-colors.
* scwm.c: Move init_image() to as late as possible to defer imlib
loading when it is used (also image-support of any kind is totally
optionally, and nothing earlier could depend on this).

* message-window.c:  Added Jeffrey Nichols to the (C), since this is
largely his code.  Renamed ResizeWindow to ResizeMessageWindow for
disambiguation.  Added `id->message-window' primitive as a last
resort in case a message window has no easy way to get a hold of
its object.  hide message windows when they are freed -- leaving
one up that's been garbage-collected would be bad.

* events.c: Replace an assert(False) with a warning about
an unhandled format in a client message.  Sam Steingold tripped
this assertion.

* color.c:  Have ScmMakeColor return BLACK_COLOR instead of
SCM_BOOL_F -- this was part of the point of my earlier reworking
of this code, but I must've gotten sidetracked.  Now, e.g.,
adjust_brightness will always get a valid color back, even if it's
not the best one.