* events.c: Fix bug that scwmdoc caught in argnumber on a VALIDATE call.
* window.c, face.c: Added use of new scwmdoc-understood /*SCWM_VALIDATE*/
comment pointing out what the argument list is.  This isn't
perfect as it's redundant with the prototype, but it's a quick fix
(and already helped me catch one bug in face.c.  Also start using
SCWM_WRONG_TYPE_ARG, instead of scm_wrong_type_arg when the
function is FUNC_NAME.