* window.h: Added tbar_right field since SetBorderX now needs it and it's computed in SetupFrame.  BE SURE TO REBUILD ALL MODULES AFTER THIS CHANGE!
* window.c: Comment an empty if conditional. Twiddle a shaded & squashed
setup-frame parameter.

* borders.h: Expose SetShapedTitlebar -- animated-window-unshade
needs it.

* events.c: ResizePswToCurrentSize when squashed titlebar and the
window title changes (since the titlebar may need adjusting).

* borders.c: Make squashed titlebars look a bit nicer-- uses the
top side as just above the client area (instead of above the
titlebar), gets the top-right corner correct, turns off relieving
of the title bar, and sizes the shaped window appropriately.  Also
relieves the top side only from the end of the title bar over to
the right.  Should be no change for non-squashed-titlebar
windows.  Fix the shaped window for squashed titlebars to not
include the window frame if the window is shaded.

* resize.c: Resize properly when grabbing the NE corner of a
squashed window.  Also round when constraining the size to the
client units, instead of truncating (better behaviour even for
non-squashed windows).