NOTE: These changes must still be made to the draw_xpm_menu.c file.
* drawmenu.c: Biggish rewrite of the menuitem flexibility stuff I
implemented too quickly yesterday.  Added HLBGColor, HLTextColor
to MenuDrawingInfo-s, drop TextColor and font.  Move TextColor and
font into MenuItemDrawinInfo struct, and have that have a BGColor
value, too.  Manage all the drawing and color setting -- this code
has too much optimization in avoiding GC manipulation; it cannot
be *that* expensive to mess with GCs.  Get the height of an item
right based on the items font size -- do not assume they'll all be
the same height.

* menu.h: Added scmHLBGColor, scmHLTextColor to menus-- does not
seem to make sense to let the highlight color change on a per item
basis.  Also added fHighlightRelief to permit turning off
relieving if the highlight colors are to be used exclusively for
indicating the selected menu item.

* menu.c: Mark new objects, provide getters/setters:
`(set-)menu-highlight-colors(!)',
`(set-)menu-highlight-relief(!/?)'

* color.h: Added SAFE_COLOR_USE_DEF and DYNAMIC_SAFE_COLOR_USE_DEF
macros since I was thinking 0 was a valid out-of-range value for
Pixels (but it's not-- it's often black! duh!)