* window.c: Fix `stick' primitive again-- need to have sticky set properly before changing the window's position. (convert_move_data) fixed bug in not returning *pStart[XY] in virtual coords.
* base.scm: Simplify and fix `move-to'.

* auto-raise.scm: Move the `provide' to the end-- causes
inexplicably problems when (use-module (app scwm auto-raise) is at
the start of my .scwmrc file if the provide is at the beginning of
this file.  Very strange, indeed!

* animation.scm: Simplify and correct `animated-move-to' -- some
of the extra junk seems to have been added (maybe by me!) as
workarounds to the bug in convert_move_data that I just fixed.

* gjb.scwmrc: Add auto-raise to the use-modules list, Added
binding for delete-window using 4 modifiers so it's hard to do by
mistake