* binding.c: Eliminate harmless compiler warning.
* miscprocs.c (s_get_key_event): Do not change the event mask for
Scr.NoFocusWin -- it already is watching for KeyRelease events, so
let it be (bug was that I was resetting the mask, thus
un-selecting events that should've been selected -- this made some
bindings appear to not work, when actually X just wasn't passing
on the event.

* virtual.c: BroadcastIconInfo of icons on other desks so that
they don't appear to move (since the Fvwm modues use viewport
positions, they need to be told when the viewport location of an
icon changes, which corresponds to when the viewport-offset
changes). Also simplify handling of focus in changeDesks -- just
call CoerceEnterNotifyOnCurrentWindow, and let the EnterNotify
event do the right thing.