* virtual.h, scwm.c, screen.h, events.c: Eliminate NON_VIRTUAL compile-time option.
* events.c: Fix bug in xterm active icons whereby the xterm would
send configure notify events that Scwm incorrectly interpreted as
referring to the client window instead of as to the pixmap icon
window (added a check to see if the event's window was
icon_pixmap_w (used to only check for icon_w)). Thanks George
Hansper for the bug report!

* binding.c: Added list of contexts to the docs for bind-key, and
xref for bind-mouse.

* virtual.c: Fix bug in initializing the scroll-region variable
that was causing edge enter events to appear to do nothing.