* window.c (select_viewport_position,get_window): Use VALIDATE_ARG_CURSOR_COPY_USE_KILLORCIRCLE to do the right thing with values #t and #f for the cursor (was getting run-time errors about uses of #t for the kill cursor)
* scwm.c: Change the SigDoneSegv message to suggest running with
--segv-just-stop or --segv-reset-count 0, and call scwm_msg only
once instead of once per line.

* resize.c: Permit rubberband mask in [0,65535], not just [0,255]

* menuitem.h, menu.h: Added self field to Menu and MenuItem
structures, a pointer back to the scheme object. Need this to be
sure we can protect/unprotect the object so that they don't get
gc-d when we're still using the guts of the objects from
DynamicMenu and MenuItemInMenu instances.

* menu.c: Initialize self fields, and use
scm_protect_object/scm_unprotect_object to ensure menuitems and
menus don't get gc-d while the menu is displayed.

* cursor.h: Rename and fix VALIDATE_ARG_CURSOR_COPY_USE_NONE to
VALIDATE_ARG_CURSOR_COPY_USE_KILLORCIRCLE