* xmisc.h, xmisc.c (NFromXPropertyCardinal): Added this fn.
* window.c: Move sticky windows off of the viewport back onto the
viewport (but not to 0,0 like the old ifdef-fed out code did).
Warn user that we did this via scwm_msg(ERR,..).

* scwm.c, add_window.c: Use SCWM_VIEWPORT_OFFSET_X/Y Root window
properties to correct for position of windows that were not on the
viewport during a previous scwm termination (e.g., when
restarting).  Also use them to go back to that viewport at
startup.