* test-case.scm: Return answer, not #<undefined>.
* reflection.scm: Workaround (procedure-source current-module)
bug.  Fix `procedure-required-formals' to not consider a rest args
to be a required argument (it kind of is, but can get bound to an
empty list). Added `apropos-internal-with-modules',
`procedure-is-interactive?', etc.

* prompt-range.scm: Tweak some of the look.

* prompt-proc.scm: Much, much improved version that operates as a
primitive documentation browser, too.  Still could use lots of
work, of course.

* preferences.scm: Put ok, apply, cancel buttons all on the same
row and make them look better.  No more stupid "Done and dismiss"
button!

* constraints.scm: Make the solver variable public at the
top-level so we can get at it.  (Changed because of a possible
quirky behaviour--- might be wholly unnecessary, but could be
useful anyway, so I'm keeping it).

* minimal.scm, base.scm: Move hook compatibility stuff from
base.scm into minimal.scm.