* window.c (s_move_window_to_desk): Return immediately if moving window to the desk it is currently on.
* placement.c: Added some debug code.  Call
MovePswToCurrentPosition() after interactive placement to
eliminate the jumping of the mapped window from 0,0 to its final
place.

* icons.c: ws change, drop some unused code that was #if 0.

* add_window.c: ws change.

* gnome-hints.scm: Added some commented out debugging.  Do not set
a window to be maximized when it is also set to be minimized (gmc
desktop_icons are this way, and maximizing them was causing them
to lose their y positioin -- all went to (0,0) in error).

* system.scwmrc: Added comment re: gmc icons.