* placement.c: Consistently subtract (psw->old_bw + psw->bw) from the final x/y positions for the placement functions.  The various fns were all different w.r.t. whether they did this.
* events.c: Added `X-SelectionNotify-hook', and handle
SelectionNotify events by calling the hook.  Useful with new
`X-convert-selection' primitive.

* xproperty.c: Added `X-convert-selection',
`X-get-selection-owner'.  Make `X-property-delete!' take an Atom
or a String, instead of requiring a string.

* window.h: Fix aliasing of "w" variable bug in
VALIDATE_ARG_WIN_ROOTSYM_OR_NUM_COPY.

* window.c (SetScwmWindowGeometry): Correct x/y after using
ConstrainSize to account for gravity.  No remarkable perceived
effect:  I'm trying to eliminate the jumpiness of the right edge
of an East-gravity window when resizing it out to the left (so
that it's right edge should stay exactly where it is throughout
the animation)

* resize.h, resize.c (ComputeDeltaForResizeWithOrigSize): Added
this function, and rewrote ComputeDeltaForResize in terms of it.
This permits using something other than the current
FRAME_WIDTH/FRAME_HEIGHT for the size (as was needed by
c-animation.c).