* xproperty.c: Fix #define FUNC_NAME mismatch caught by scwmdoc.
* events.c (HandleKeyEvent): Ungrab the keyboard when we're about
to do the execution of the commands.  This lets C-S-M-Enter, e.g.,
still be sent to the application (XTerm, e.g.) even when C-S-M is
a three-modifier binding that would otherwise wait for the Meta to
be released before releasing the passive grab.  Fixs bug reported
by Stefan Saroui re: C-S-M-Enter in XTerms no longer working since
I added the fvwm2-pager auto-raising on C-S-M modifier binding.

* binding.c: Let `unbind-keycode' take two ignored-proc arguments
to permit calls that are otherwise identical to `bind-keycode'